iot-kafka-monitoring/requirements.txt
kafka-python==2.0.2   # Biblioteca cliente para conectar, produzir e consumir mensagens do Apache Kafka em Python.
faker==24.3.0         # Gera dados falsos e aleatórios (nomes, datas, números, etc.), útil para simular sensores IoT.
psycopg2-binary==2.9.10
redis==6.2.0